Add files via upload
Beauty parlor is a service-oriented establishment in which women receive treatment to increase their beauty. Skincare, facial make-up, hair nourishment, and model hair cutting are the most important nourishing activities of a beauty parlor. In a parlor, there is a manager who manages all the documents with details. Using pen & paper for collecting documents isn’t safe and secure.

Hence, a management system is introduced to maintain a parlor record. It’s necessary to have a complete project on it so that the documents are safe and they can redo them whenever they want. 

It is a database management system which is helpful in the sectors of beauty parlor.
Our main aim is to maintain records and make analyses based on reports generated. 

The system can store details of the appointment, employee, payment info, stock details, etc. 

This system is built to reduce the complexity of the system for the users handling the system.